put gene test in a separate file
